quantnity uses new API
