+++ autogalaxy/quantity/model/analysis.py
@@ -1,5 +1,3 @@
-import os
-
 from autoconf.dictable import to_dict
 
 import autofit as af
@@ -7,9 +5,9 @@ import autofit as af
 from autogalaxy.analysis.analysis.analysis import Analysis
 from autogalaxy.cosmology.lensing import LensingCosmology
 from autogalaxy.cosmology.wrap import Planck15
-from autogalaxy.quantity.model.plotter_interface import PlotterInterfaceQuantity
 from autogalaxy.quantity.dataset_quantity import DatasetQuantity
 from autogalaxy.quantity.model.result import ResultQuantity
+from autogalaxy.quantity.model.visualizer import VisualizerQuantity
 from autogalaxy.quantity.fit_quantity import FitQuantity
 
 from autogalaxy import exc
@@ -17,6 +15,7 @@ from autogalaxy import exc
 
 class AnalysisQuantity(Analysis):
     Result = ResultQuantity
+    Visualizer = VisualizerQuantity
 
     def __init__(
         self,
@@ -121,47 +120,6 @@ class AnalysisQuantity(Analysis):
         return FitQuantity(
             dataset=self.dataset, light_mass_obj=galaxies, func_str=self.func_str
         )
-
-    def visualize(
-        self,
-        paths: af.DirectoryPaths,
-        instance: af.ModelInstance,
-        during_analysis: bool,
-    ):
-        """
-        Output images of the maximum log likelihood model inferred by the model-fit. This function is called throughout
-        the non-linear search at regular intervals, and therefore provides on-the-fly visualization of how well the
-        model-fit is going.
-
-        The visualization performed by this function includes:
-
-        - Images of the best-fit galaxy.
-
-        - Images of the best-fit `FitQuantity`, including the model-image, residuals and chi-squared of its fit to
-        the imaging data.
-
-        The images output by this function are customized using the file `config/visualize/plots.yaml`.
-
-        Parameters
-        ----------
-        paths
-            The PyAutoFit paths object which manages all paths, e.g. where the non-linear search outputs are stored,
-            visualization, and the pickled objects used by the aggregator output by this function.
-        instance
-            An instance of the model that is being fitted to the data by this analysis (whose parameters have been set
-            via a non-linear search).
-        during_analysis
-            If True the visualization is being performed midway through the non-linear search before it is finished,
-            which may change which images are output.
-        """
-
-        if os.environ.get("PYAUTOFIT_TEST_MODE") == "1":
-            return
-
-        fit = self.fit_quantity_for_instance(instance=instance)
-
-        PlotterInterface = PlotterInterfaceQuantity(output_path=paths.image_path)
-        PlotterInterface.fit_quantity(fit=fit)
 
     def save_attributes(self, paths: af.DirectoryPaths):
         """
