+++ test_autogalaxy/interferometer/model/test_plotter_interface_interferometer.py
@@ -1,7 +1,9 @@
 from os import path
 import pytest
 
-from autogalaxy.interferometer.model.plotter_interface import PlotterInterfaceInterferometer
+from autogalaxy.interferometer.model.plotter_interface import (
+    PlotterInterfaceInterferometer,
+)
 
 directory = path.dirname(path.abspath(__file__))
 
@@ -11,9 +13,7 @@ def make_plotter_interface_plotter_setup():
     return path.join("{}".format(directory), "files")
 
 
-def test__interferometer(
-    interferometer_7, include_2d_all, plot_path, plot_patch
-):
+def test__interferometer(interferometer_7, include_2d_all, plot_path, plot_patch):
     plotter_interface = PlotterInterfaceInterferometer(output_path=plot_path)
 
     plotter_interface.interferometer(dataset=interferometer_7)
